move library to better location
